requirements.txt
aniso8601==9.0.1
bidict==0.21.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
cryptography==3.4.7
Flask==2.0.1
Flask-RESTful==0.3.9
Flask-SocketIO==5.1.0
Flask-SQLAlchemy==2.5.1
gevent==21.1.2
greenlet==1.1.0
idna==3.2
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycparser==2.20
python-dotenv==0.19.0
python-engineio==4.2.0
python-socketio==5.3.0
pytz==2021.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.22
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
zipp==3.5.0
zope.event==4.5.0
zope.interface==5.4.0
